requirements.txt
flask==3.1.0
requests==2.32.3
spidev==3.6
pyserial==3.5